requirements.txt
pytest==7.2.*
pytest-cov==4.0.*
pytest-flask==1.2.*
psycopg2-binary==2.9.*
requests==2.31.0
black==23.10.1
pre-commit==3.5.0
google-auth==2.18.*
Flask-Cors==3.0.*
Flask-JWT-Extended==4.4.*
flask-smorest==0.39.*
Flask==2.2.*
SQLAlchemy==2.0.*
alembic==1.10.*
dynaconf==3.1.*
sqlalchemy-easy-profile==1.3.*
sqlalchemy-utils==0.40.*
typing-extensions==4.4.*
qrcode==7.4.*
pyshorteners==1.0.*
